Move stuck split test to a separate class
The test mixed class-level state with method-level
state and wasn't consistent with everything that was
going on in that class.
